hi-spring/src/test/java/yungdo/hispring/repository/MemoryMemberRepositoryTest.java
package yungdo.hispring.repository;

import org.junit.jupiter.api.AfterEach;
import yungdo.hispring.domain.Member;
import org.junit.jupiter.api.Test;

import java.util.List;
import java.util.Optional;

import static org.assertj.core.api.Assertions.*;

class MemoryMemberRepositoryTest {

    MemoryMemberRepository repository = new MemoryMemberRepository();

    @AfterEach //메서드가 실행이 끝날때마다 실행되는 콜백 메서드, 테스트가 실행될때마다 store가 비워지면서 테스트 순서가 상관없게 됨(의존성이 없게 됨)
    public void afterEach(){
        repository.clearStore();
    }

    @Test
    public void save(){
        Member member = new Member();
        member.setName("Spring");

        repository.save(member);

        Member result = repository.findById(member.getId()).get();
        assertThat(member).isEqualTo(result); // result랑 member랑 똑같은지 검증
    }

    @Test
    public void findByName() {
        Member member1 = new Member();
        member1.setName("Spring1");
        repository.save(member1);

        Member member2 = new Member();
        member2.setName("Spring1");
        repository.save(member2);

        Member result = repository.findByName("Spring1").get(); //get하면 까서 나옴 ,,? spring2로 바꾸면 오류남

        assertThat(result).isEqualTo(member1);
    }

    @Test
    public void findAll() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.save(member1);

        Member member2 = new Member();
        member2.setName("spring2");
        repository.save(member2);

        List<Member> result = repository.findAll();

        assertThat(result.size()).isEqualTo(2);

    }

}
